python_deprecated/requirements.txt
google-genai
discord.py
python-dotenv